+++ src/lib/components/nodes/ElementaryConstructor.svelte
@@ -19,7 +19,7 @@
     type Node,
   } from '@xyflow/svelte'
   import { nodeColors, NodeType, Type } from '../../types/nodeTypes'
-  import { getImportedNodesByType } from '../../states/store.svelte'
+  import { getImportedNodesByType, removeNode } from '../../states/store.svelte'
 
   let { id, data, type }: NodeProps<ElementaryConstructorType> = $props()
 
@@ -53,7 +53,10 @@
 </script>
 
 <div class="custom-node" style="--border-color: {color}">
-  <div class="label">{data.type}</div>
+  <div class="node-header">
+    <div class="label">{data.type}</div>
+    <button class="button-remove" onclick={() => removeNode(id)}>X</button>
+  </div>
   <div>
     {#if type === Type.UNSIGNED}
       <input
@@ -108,8 +111,25 @@
     border: 2px solid var(--border-color);
   }
 
+  .node-header {
+    display: flex;
+    justify-content: space-between;
+    gap: 1vw;
+  }
+
   .label {
     font-weight: bold;
+  }
+
+  .button-remove {
+    cursor: pointer;
+    border: 1px solid var(--border-color);
+    border-radius: 3px;
+    margin: 0 0 1vh 1vh;
+  }
+
+  .button-remove:hover {
+    border: 1px solid var(--border-color-hover);
   }
 
   input.invalid {