+++ src/lib/components/nodes/EmptyConstructor.svelte
@@ -8,9 +8,9 @@
 <script lang="ts">
   import { Handle, Position, type NodeProps, type Node } from '@xyflow/svelte'
   import { nodeColors, NodeType, Type } from '../../types/nodeTypes'
-  import { getImportedNodesByType } from '../../states/store.svelte'
+  import { getImportedNodesByType, removeNode } from '../../states/store.svelte'
 
-  let { data, type }: NodeProps<EmptyConstructorType> = $props()
+  let { id, data, type }: NodeProps<EmptyConstructorType> = $props()
 
   const importedNodes = getImportedNodesByType(NodeType.EMPTY_CONSTRUCTOR)
   const color = nodeColors[NodeType.EMPTY_CONSTRUCTOR]
@@ -34,7 +34,10 @@
 </script>
 
 <div class="custom-node" style="--border-color: {color}">
-  <div class="label">{data.type}</div>
+  <div class="node-header">
+    <div class="label">{data.type}</div>
+    <button class="button-remove" onclick={() => removeNode(id)}>X</button>
+  </div>
   {#each data.outputs as i, index (i)}
     <Handle id={`output-${index}`} type="source" position={Position.Right} />
   {/each}
@@ -46,6 +49,23 @@
     border-radius: 5px;
     background: var(--secondary-color);
     border: 2px solid var(--border-color);
+  }
+
+  .node-header {
+    display: flex;
+    justify-content: space-between;
+    gap: 1vw;
+  }
+
+  .button-remove {
+    cursor: pointer;
+    border: 1px solid var(--border-color);
+    border-radius: 3px;
+    margin: 0 0 1vh 1vh;
+  }
+
+  .button-remove:hover {
+    border: 1px solid var(--border-color-hover);
   }
 
   .label {